+++ lib/augen.dart
@@ -8,5 +8,6 @@ export 'src/models/ar_node.dart' show ARNode, NodeType, ModelFormat;
 export 'src/models/ar_plane.dart';
 export 'src/models/ar_hit_result.dart';
 export 'src/models/ar_session_config.dart';
+export 'src/models/ar_animation.dart';
 export 'src/models/vector3.dart';
 export 'src/models/quaternion.dart';